designToCodeAiAgent/requirements.txt
# ==========================================
# Figma to MiBlock CMS Component Generator
# Python Dependencies (Python 3.13 compatible)
# ==========================================

# Web Framework
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.4.1
python-socketio==5.11.4
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.11
pgvector==0.3.6
sqlalchemy==2.0.36
alembic==1.14.0

# Redis Cache - REMOVED (using in-memory cache instead for simplicity)
# redis==5.2.1
# hiredis==3.0.0

# AI/ML (Python 3.13 compatible versions)
anthropic==0.42.0
openai==1.58.1
torch>=2.5.0
torchvision>=0.20.0
transformers==4.47.1
pillow==11.0.0
sentence-transformers==3.3.1  # For embeddings (includes CLIP models)

# Figma API
requests==2.32.3
httpx==0.28.1

# Image Processing & Comparison
opencv-python==4.10.0.84
scikit-image==0.24.0
imagehash==4.3.1
numpy==2.2.1

# HTML Processing
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1

# Agent Orchestration
langgraph==0.2.59
langchain==0.3.13
langchain-core==0.3.28
langchain-anthropic==0.3.0

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Monitoring & Logging
structlog==24.4.0
python-json-logger==3.2.1

# Development Tools
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# Additional Utilities
tabulate==0.9.0  # For database check script
